WIP: Work on automatic group addition for users
If the host machine has a program/feature, that requires group access,
enabled the user gets added to the group automatically.
Since some groups can cause problems for other users (priviledge) these
are then combed for priviledged and default. If the user is not
priviledged the groups, even if the programs are installed (groups
created) is not part of it.