Change wallpaper, again.
Actually nice resolution.
No idea who the author is (RIP).
The red, even the lighter parts, are not straining on the eyes in the
evening the same way previous one was.
Even text visibility (white) is basically amazing, I don't see much
loss, even when the window is unfocused. Damn cool.
Thinking of making per-host or per-user wallpaper. Both are easy to do
either way.